requirements.txt
langchain
langchain-openai
langchain-community
langchain-classic
faiss-cpu
streamlit
python-dotenv
sentence-transformers